render the header once in an (app) layout
Every page rendered its own SharedHeader, so each navigation unmounted
the whole tree and the header flashed. The authenticated routes now live
in an (app) route group whose layout renders the header and tab bar once;
only the content below swaps on client-side navigation.

Pages hand their section anchors to the header through a small context
instead of a bottomBar prop, and the cart badge is fed by the shared
badge store (CartProvider and the cart page write to it, the header
refreshes it on every route change) instead of a cartBadgeOverride prop.
